feat: make main content server side generated
